# src/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This api call data from youtube anf table",
        "title": "Dataset Collection API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "10.201.252.7:30014",
    "paths": {
...
        "/dataset/queryTrainLprTag/{youtubeId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lpr resualt"
                ],
                "summary": "trigger video into lpr detect image",
                "operationId": "query-train-lpr-tag-by-youtubeId-handler",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Youtube ID",
                        "name": "youtubeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "We need Youtube ID!!",
                        "schema": {
                            "$ref": "#/definitions/string"
                        }
                    }
                }
            }
        },
...
        "/dataset/queryTrainYoloTag/{youtubeId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "yolo resualt"
                ],
                "summary": "get yolo detect info by youtubeId",
                "operationId": "query-train-yolo-tag-by-youtubeId-handler",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Youtube ID",
                        "name": "youtubeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "We need Youtube ID!!",
                        "schema": {
                            "$ref": "#/definitions/string"
                        }
                    }
                }
            }
        },
...
        "/dataset/youtubeUrl/cartype/{youtubeId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dataset car type"
                ],
                "summary": "download video by youtubeId",
                "operationId": "url-to-download-car-type",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Youtube ID",
                        "name": "youtubeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "We need youtube Id!!",
                        "schema": {
                            "$ref": "#/definitions/string"
                        }
                    }
                }
            }
        },
...
    }
}